Refactor login to a separate task
