+++ src/main/java/org/jenkinsci/plugins/spoontrigger/ScriptBuilder.java
@@ -1,9 +1,6 @@
 package org.jenkinsci.plugins.spoontrigger;
 
-import com.cloudbees.plugins.credentials.common.StandardListBoxModel;
-import com.cloudbees.plugins.credentials.common.StandardUsernamePasswordCredentials;
 import com.google.common.base.Optional;
-import com.google.common.base.Strings;
 import com.google.common.reflect.TypeToken;
 import hudson.*;
 import hudson.model.*;
@@ -15,8 +12,6 @@ import lombok.Data;
 import lombok.Getter;
 import org.jenkinsci.plugins.spoontrigger.commands.CommandDriver;
 import org.jenkinsci.plugins.spoontrigger.commands.turbo.BuildCommand;
-import org.jenkinsci.plugins.spoontrigger.commands.turbo.ConfigCommand;
-import org.jenkinsci.plugins.spoontrigger.commands.turbo.LoginCommand;
 import org.jenkinsci.plugins.spoontrigger.commands.turbo.VersionCommand;
 import org.jenkinsci.plugins.spoontrigger.hub.Image;
 import org.jenkinsci.plugins.spoontrigger.utils.AutoCompletion;
@@ -30,23 +25,16 @@ import org.kohsuke.stapler.QueryParameter;
 import javax.annotation.Nullable;
 import java.io.File;
 import java.io.IOException;
-import java.util.List;
 
 import static com.google.common.base.Preconditions.checkState;
 import static org.jenkinsci.plugins.spoontrigger.Messages.*;
 import static org.jenkinsci.plugins.spoontrigger.utils.LogUtils.log;
 
-public class ScriptBuilder extends BaseBuilder {
+public class ScriptBuilder extends LoginBuilder {
 
     @Nullable
     @Getter
     private final String scriptFilePath;
-    @Nullable
-    @Getter
-    private final String credentialsId;
-    @Nullable
-    @Getter
-    private final String hubUrl;
     @Nullable
     @Getter
     private final String imageName;
@@ -72,9 +60,9 @@ public class ScriptBuilder extends BaseBuilder {
     public ScriptBuilder(String scriptFilePath, String credentialsId, String hubUrl, String imageName,
                          String vmVersion, String containerWorkingDir, @Nullable MountSettings mountSettings,
                          boolean noBase, boolean overwrite, boolean diagnostic) {
+        super(credentialsId, hubUrl);
+
         this.scriptFilePath = Util.fixEmptyAndTrim(scriptFilePath);
-        this.credentialsId = Util.fixEmptyAndTrim(credentialsId);
-        this.hubUrl = Util.fixEmptyAndTrim(hubUrl);
         this.imageName = Util.fixEmptyAndTrim(imageName);
         this.vmVersion = Util.fixEmptyAndTrim(vmVersion);
         this.containerWorkingDir = Util.fixEmptyAndTrim(containerWorkingDir);
@@ -106,12 +94,10 @@ public class ScriptBuilder extends BaseBuilder {
 
     @Override
     public void prebuild(SpoonBuild build, BuildListener listener) {
+        super.prebuild(build, listener);
+
         checkState(build.getEnv().isPresent(), "Env is not defined");
 
-        Optional<StandardUsernamePasswordCredentials> credentials = this.getCredentials();
-        if (credentials.isPresent()) {
-            build.setCredentials(credentials.get());
-        }
         build.setAllowOverwrite(this.overwrite);
 
         FilePath scriptPath = this.resolveScriptFilePath(build, build.getEnv().get(), listener);
@@ -122,16 +108,13 @@ public class ScriptBuilder extends BaseBuilder {
 
     @Override
     public boolean perform(SpoonBuild build, Launcher launcher, BuildListener listener) throws IOException, InterruptedException {
+        if (!super.perform(build, launcher, listener)) {
+            return false;
+        }
+
         CommandDriver client = CommandDriver.scriptBuilder(build).launcher(launcher).listener(listener).ignoreErrorCode(true).build();
 
         checkSpoonPluginIsRunning(client);
-
-        switchHub(client);
-
-        Optional<StandardUsernamePasswordCredentials> credentials = build.getCredentials();
-        if (credentials.isPresent()) {
-            login(client, credentials.get());
-        }
 
         BuildCommand command = createBuildCommand(build.getScript().get());
         command.run(client);
@@ -169,23 +152,6 @@ public class ScriptBuilder extends BaseBuilder {
         versionCmd.run(client);
     }
 
-    private void switchHub(CommandDriver client) {
-        ConfigCommand.CommandBuilder cmdBuilder = ConfigCommand.builder();
-        if (Strings.isNullOrEmpty(this.hubUrl)) {
-            cmdBuilder.reset(true);
-        } else {
-            cmdBuilder.hub(this.hubUrl);
-        }
-
-        ConfigCommand configCommand = cmdBuilder.build();
-        configCommand.run(client);
-    }
-
-    private void login(CommandDriver client, StandardUsernamePasswordCredentials credentials) {
-        LoginCommand loginCmd = LoginCommand.builder().login(credentials.getUsername()).password(credentials.getPassword()).build();
-        loginCmd.run(client);
-    }
-
     private BuildCommand createBuildCommand(FilePath scriptPath) {
         BuildCommand.CommandBuilder cmdBuilder = BuildCommand.builder().script(scriptPath);
         if (this.imageName != null) {
@@ -209,18 +175,6 @@ public class ScriptBuilder extends BaseBuilder {
         cmdBuilder.noBase(this.noBase);
 
         return cmdBuilder.build();
-    }
-
-    private Optional<StandardUsernamePasswordCredentials> getCredentials() throws IllegalStateException {
-        if (Strings.isNullOrEmpty(this.credentialsId)) {
-            return Optional.absent();
-        }
-
-        Optional<StandardUsernamePasswordCredentials> credentials = Credentials.lookupById(StandardUsernamePasswordCredentials.class, this.credentialsId);
-
-        checkState(credentials.isPresent(), "Cannot find any credentials with id (%s)", this.credentialsId);
-
-        return credentials;
     }
 
     private FilePath resolveScriptFilePath(AbstractBuild build, EnvVars environment, BuildListener listener) throws IllegalStateException {
@@ -288,7 +242,6 @@ public class ScriptBuilder extends BaseBuilder {
         private static final Validator<String> SCRIPT_FILE_PATH_STRING_VALIDATOR;
         private static final Validator<File> SCRIPT_FILE_PATH_FILE_VALIDATOR;
         private static final Validator<String> VERSION_NUMBER_VALIDATOR;
-        private static final Validator<String> CREDENTIALS_ID_VALIDATOR;
         private static final Validator<String> NULL_OR_SINGLE_WORD_VALIDATOR;
 
         static {
@@ -302,7 +255,6 @@ public class ScriptBuilder extends BaseBuilder {
             VERSION_NUMBER_VALIDATOR = Validators.chain(
                     IGNORE_NULL_VALIDATOR,
                     StringValidators.isVersionNumber());
-            CREDENTIALS_ID_VALIDATOR = StringValidators.isNotNull("Credentials are required to login to a Turbo account", Level.WARNING);
             NULL_OR_SINGLE_WORD_VALIDATOR = Validators.chain(
                     IGNORE_NULL_VALIDATOR,
                     StringValidators.isSingleWord(String.format(REQUIRE_SINGLE_WORD_S, "Parameter")));
@@ -343,16 +295,6 @@ public class ScriptBuilder extends BaseBuilder {
             return Validators.validate(IGNORE_NULL_VALIDATOR, workingDirectory);
         }
 
-        public FormValidation doCheckCredentialsId(@AncestorInPath Item project, @QueryParameter String value) {
-            if (doNotHasPermissions(project)) {
-                return FormValidation.ok();
-            }
-
-            String credentialsId = Util.fixEmptyAndTrim(value);
-            Validator<String> validator = Validators.chain(CREDENTIALS_ID_VALIDATOR, new CredentialValidator(project));
-            return Validators.validate(validator, credentialsId);
-        }
-
         public FormValidation doCheckImageName(@QueryParameter String value) {
             String imageName = Util.fixEmptyAndTrim(value);
             return Validators.validate(NULL_OR_SINGLE_WORD_VALIDATOR, imageName);
@@ -368,13 +310,12 @@ public class ScriptBuilder extends BaseBuilder {
             return Validators.validate(SCRIPT_FILE_PATH_STRING_VALIDATOR, sourceFolder);
         }
 
-        public ListBoxModel doFillCredentialsIdItems(@AncestorInPath Item project) {
-            if (doNotHasPermissions(project)) {
-                return new StandardListBoxModel();
-            }
+        public FormValidation doCheckCredentialsId(@AncestorInPath Item project, @QueryParameter String value) {
+            return Credentials.checkCredetntials(project, value);
+        }
 
-            List<StandardUsernamePasswordCredentials> projectCredentials = Credentials.lookupByItem(StandardUsernamePasswordCredentials.class, project);
-            return new StandardListBoxModel().withEmptySelection().withAll(projectCredentials);
+        public ListBoxModel doFillCredentialsIdItems(@AncestorInPath Item project) {
+            return Credentials.fillCredentialsIdItems(project);
         }
 
         @Override
@@ -385,28 +326,6 @@ public class ScriptBuilder extends BaseBuilder {
         @Override
         public String getDisplayName() {
             return "Execute TurboScript";
-        }
-
-        private static final class CredentialValidator implements Validator<String> {
-
-            private final Item project;
-
-            public CredentialValidator(Item project) {
-                this.project = project;
-            }
-
-            @Override
-            public void validate(String credentialsId) throws ValidationException {
-                Optional<StandardUsernamePasswordCredentials> credentials = Credentials.lookupById(StandardUsernamePasswordCredentials.class, project, credentialsId);
-
-                if (credentials.isPresent()) {
-                    return;
-                }
-
-                String errMsg = String.format("Cannot find any credentials with id (%s)", credentialsId);
-                FormValidation formValidation = FormValidation.warning(errMsg);
-                throw new ValidationException(formValidation);
-            }
         }
     }
 }
