+++ src/main/java/org/jenkinsci/plugins/spoontrigger/utils/Credentials.java
@@ -2,19 +2,28 @@ package org.jenkinsci.plugins.spoontrigger.utils;
 
 import com.cloudbees.plugins.credentials.CredentialsMatchers;
 import com.cloudbees.plugins.credentials.CredentialsProvider;
+import com.cloudbees.plugins.credentials.common.StandardListBoxModel;
+import com.cloudbees.plugins.credentials.common.StandardUsernamePasswordCredentials;
 import com.cloudbees.plugins.credentials.common.UsernameCredentials;
 import com.cloudbees.plugins.credentials.domains.DomainRequirement;
 import com.google.common.base.Optional;
+import hudson.Util;
 import hudson.model.Item;
 import hudson.model.ItemGroup;
 import hudson.security.ACL;
+import hudson.util.FormValidation;
+import hudson.util.ListBoxModel;
 import jenkins.model.Jenkins;
+import org.jenkinsci.plugins.spoontrigger.validation.*;
+import org.kohsuke.stapler.AncestorInPath;
+import org.kohsuke.stapler.QueryParameter;
 
 import java.util.List;
 
 public final class Credentials {
 
     private static final DomainRequirement ANY_DOMAIN = new DomainRequirement();
+    private static final Validator<String> CREDENTIALS_ID_VALIDATOR = StringValidators.isNotNull("Credentials are required to login to a Turbo account", Level.WARNING);
 
     public static <T extends UsernameCredentials> Optional<T> lookupById(Class<T> aClass, String credentialId) {
         List<T> availableCredentials = lookupByItemGroup(aClass, Jenkins.getInstance());
@@ -32,5 +41,28 @@ public final class Credentials {
 
     private static <T extends UsernameCredentials> List<T> lookupByItemGroup(Class<T> aClass, ItemGroup itemGroup) {
         return CredentialsProvider.lookupCredentials(aClass, itemGroup, ACL.SYSTEM, ANY_DOMAIN);
+    }
+
+    public static ListBoxModel fillCredentialsIdItems(@AncestorInPath Item project) {
+        if (doNotHasPermissions(project)) {
+            return new StandardListBoxModel();
+        }
+
+        List<StandardUsernamePasswordCredentials> projectCredentials = Credentials.lookupByItem(StandardUsernamePasswordCredentials.class, project);
+        return new StandardListBoxModel().withEmptySelection().withAll(projectCredentials);
+    }
+
+    public static FormValidation checkCredetntials(@AncestorInPath Item project, @QueryParameter String value) {
+        if (doNotHasPermissions(project)) {
+            return FormValidation.ok();
+        }
+
+        String credentialsId = Util.fixEmptyAndTrim(value);
+        Validator<String> validator = Validators.chain(CREDENTIALS_ID_VALIDATOR, new CredentialValidator(project));
+        return Validators.validate(validator, credentialsId);
+    }
+
+    private static boolean doNotHasPermissions(Item project) {
+        return project == null || !project.hasPermission(Item.CONFIGURE);
     }
 }
